Create zigzagtraversal.cpp
Zig-zag traversal of tree

In this kind of traversal we do one level in the way of left to right and the below level we do right to left.
Time Complexity: O(n) as we are traversing all the nodes of the Binary tree in ZigZag traversal of Binary Tree at least once. 
Space Complexity: O(n) as extra space for storing the nodes of alternate levels in the stack has been used.